requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pylint
radon